fix punctuation
Bob DuCharme's 6th comment
https://www.w3.org/mid/33e6564c-0dc3-4ca6-8304-4e47da867bf6@snee.com

Bob points out that what looks like an enumeration of 3 things
is actually an enumeration of 2 things.
(this is actually closer to the original phrasing in RDF 1.1 Concepts,
which was "fixed" probably because of an ugly repetition of "and".
I avoided the repetition by using "then")