export all normatively defined terms
I broke my own rule of systematically exporting normative definitions...
I strongly felt that the following terms were defined for internal purposes only,
more precisely to help define other terms, so I chose to not export them.

- IRI equality
- Blank node equality
- Literal equality
- Triple equality
- isomorphic RDF-term mapping

I experimented with scoped definition [1]
for genetic terms such as 'atomic', 'node', 'subject'...
but I found that the ergonomic toll was not worth it.
In particular, there is no hint in the HTML
that the definition requires some additional context to be used.

[1] https://respec.org/docs/#scoped-concepts